==> 02c_Load_and_combine_timeseries_data.ipynb <==
# Load and combine timeseries data 

**Index**
1. [Timeseries-Estonia](#Timeseries-Estonia)
2. [Timeseries Estonian Counties](#Timeseries-Estonian-Counties)

This Notebooks shows how to load previously generated Estonian and County-level timeseries into a File Geodatabase. The data that is used, is provided by the [Maa-amet geoportaal](https://geoportaal.maaamet.ee/eng/), [Eesti Statistika](https://www.stat.ee/en) and [Terviseamet via DigiLugu](https://www.terviseamet.ee/et/koroonaviirus/avaandmed).

import arcpy
import os
from aglearn import remap as rm
from pprint import pprint

Set variables

os.getcwd()

path = os.path.join(os.getcwd(), r'gisdata')
gdbPath = os.path.join(path, r'covid.gdb')
tmpPath = os.path.join(path, r'temp.gdb')
arcpy.env.overwriteOutput = True
arcpy.env.workspace = gdbPath

## Timeseries Estonia

Here the timeseries that was created from the raw open data ([see here](01_Download_and_Preprocess_Data.ipynb#Timeseries-Estonia)), is loaded into the File Geodatabase.

eetstable = "timeseries_estonia"
result = arcpy.conversion.TableToTable(r"data/cov_ts_eesti.csv", gdbPath, eetstable)

The conversion tool automatically assumes that the numbers should be treated as float values. If this is not desired [```FieldMappings```](https://desktop.arcgis.com/en/arcmap/10.3/analyze/arcpy-classes/fieldmappings.htm) need to be created beforehand.

### Add demographic information

#### Get total population based on Ruutkaart2018

The data is stored in the point and polygon shapefiles. The arcpy cursor can be utilized to load the data. (Alternatively the [SummaryStatistics](https://pro.arcgis.com/en/pro-app/tool-reference/analysis/summary-statistics.htm) tool could be used.)

mkrkfn = r"maakond_ruutkaart_"

Loop through the all counties in Feature class, calculate total sum for population (```ePopTotal2018```) and fill 2 dictionaries:

- ```eePopCounty2018``` Population per County, MKOOD used as the key
- ``` eeShapeCounty``` Dictionary in which the geometries of each County (ArcObjects) are stored.

eePopTotal2018 = 0
eePopCounty2018 = {}
eeShapeCounty = {}
with arcpy.da.SearchCursor(mkrkfn, ["MKOOD", "TOTAL", "SHAPE@"]) as sCursor:
    for row in sCursor:
        mkood = row[0]
        pop = row[1]
        shape = row[2]
        eePopTotal2018 = int(eePopTotal2018 + pop)
        eePopCounty2018[mkood] = int(pop)
        eeShapeCounty[mkood] = shape

print('Total population of Estonia, according to underlying dataset was {}'.format(int(eePopTotal2018)))
print('It divides into the respective Counties , as follows:')
for mkood in eePopCounty2018.keys():
    print('{:<20} : {}'.format(rm.MKOOD_MNIMI[mkood], eePopCounty2018[mkood] )) # look up the County name from aglearn module

#### Calculate cases relative to Population

Using the total population it is now possible to set the number of positive cases into perspective.

result = arcpy.AddField_management(eetstable, 'cumulativePosPer10K', 'Double')
result = arcpy.CalculateField_management(eetstable, 
                                "cumulativePosPer10K",
                                "round((!cumulativePositive! / {eePop})*10000,1)".format(eePop=eePopTotal2018),
                                "PYTHON3")
result = arcpy.AddField_management(eetstable, 'testedPerfPerPop10K', 'Double')
result = arcpy.CalculateField_management(eetstable, 
                                "testedPerfPerPop10K", 
                                "round((!testsPerformed! / {eePop})*10000,1)".format(eePop=eePopTotal2018), 
                                "PYTHON3")

ts_fields = arcpy.ListFields(eetstable)
ts_fns = [] # saving the fieldnames here
print("{:<25} {:<10} {:<7}".format('Fieldname:', 'Fieldtype:', 'Length:'))
for field in ts_fields:
    ts_fns.append(field.name)
    print("{:<25} {:<10} {:<7}".format(field.name, field.type, field.length))
ts_fns.remove('OBJECTID') # no need for OBJECTID later

## Timeseries Estonian Counties

For the comprehensive dataset of Estonia a table without spatial data was sufficient. The county data shall be displayed on a map, hence a Polygon Feature Class is needed.

mktsfs = "timeseries_county"

### Create empty FeatureClass

Copying the SpatialReference from the summarized dataset created before (click here to jump back: [full](02a_Download_and_Analyze_Demographic_Data.ipynb#county_full) and [simplified](02b_Create_Simplified_Polygons.ipynb#county_simple) counties).

sr = arcpy.Describe(mkrkfn).spatialReference

result = arcpy.CreateFeatureclass_management(gdbPath, mktsfs, "POLYGON", "", "", "", sr)

Adding all the fields that are present in the previos timeseries table, plus two fields identifying the counties.

for field in ts_fields[1:]:
    arcpy.AddField_management(mktsfs, field.name, field.type)
    
arcpy.AddField_management(mktsfs, 'MKOOD', 'TEXT')
ts_fns.append('MKOOD')
arcpy.AddField_management(mktsfs, 'County', 'TEXT')
ts_fns.append('County')

### Fill FeatureClass

Pandas is used to read the CSV file and fill the FeatureClass.

(Alternative workflow: ```TableToTable``` + [```AddJoin```](https://pro.arcgis.com/en/pro-app/tool-reference/data-management/add-join.htm) + [```CopyFeature```](https://pro.arcgis.com/en/pro-app/tool-reference/data-management/copy-features.htm) ).

import pandas as pd

Read the Dataframe

df = pd.read_csv('data/ts_maakond.csv')

# correct the MKOOD, which was falsely imported as integer.
df['MKOOD'] = df['County'].map(rm.MNIMI_MKOOD) 

# get the population from previously created dictionary
df['Population2018'] = df['MKOOD'].map(eePopCounty2018) 

# calculate two new columns (similar to previously performed CalculateField in TimeSeries estonia)
df['cumulativePosPer10K'] = round((df['cumulativePositive'] / df['Population2018'])*10000,1) #
df['testedPerfPerPop10K'] = round((df['testsPerformed'] / df['Population2018'])*10000,1)
df

Create a list of fieldnames, which occure in both datasets.

sFields = list(set(ts_fns).intersection(list(df.columns)))
sFields.append('SHAPE@')
sFields

Add new data with the ```InsertCursor```.

iCursor = arcpy.da.InsertCursor(mktsfs, sFields)
for indx,row in df[:].iterrows():
    irow = [] # row to insert (to be filled)
    for field in sFields[:-1]: # for each of the fields except 'SHAPE@'
        irow.append(row[field]) # append the value
    irow.append(eeShapeCounty[row.MKOOD]) # for the last field 'SHAPE@' get ArcObject from dictionary
    iCursor.insertRow(irow)
del iCursor
