==> 03a_publish_data_into_arcgis.com.ipynb <==
# Publish Data to ArcGIS Online

## Add layer to ArcGIS Pro

from IPython.display import Video

### Enable time and set Symbology in ArcGIS Pro

In the following clip it is demonstrated how to filter by data and how to enable time on the newly created feature class, to allow for animation and use of the timeslider object in ArcGIS Pro and ArcGIS Online.
(To access the clip, clone and access the Notebook locally, or click [here](https://github.com/MBennGit/agcovdemo/blob/master/img/03a_1_enable_timeslider.mp4?raw=true).) 

Video(r"img/03a_1_enable_timeslider.mp4", width=960)

### Add remaining layer to ArcGIS Pro

Each Feature Class, that shall be published to ArcGIS Online, is added to ArcGIS Pro. Properties and Symbologies are adapted. (click [here](https://github.com/MBennGit/agcovdemo/blob/master/img/03a_2_add_remaining_layer.mp4?raw=true) to access the clip) 

Video(r"img/03a_2_add_remaining_layer.mp4", width=960)

### Publish to ArcGIS Online

Finally the layer are selected and published to ArcGIS Online. (click [here](https://github.com/MBennGit/agcovdemo/blob/master/img/03a_3_publish_data.mp4?raw=true) to access the clip)

Video(r"img/03a_3_publish_data.mp4", width=960)
